move stand alone ell1 model to stand alone binary folder and change pint_binary_ell1 name to pulsar_binary_ell1
